[ros] ros components moved into their own package, closes #35
