+++ py_trees_ros/CHANGELOG.rst
@@ -1,0 +1,7 @@
+Changelog
+=========
+
+Forthcoming
+-----------
+* ...
+
