+++ py_trees_ros/CMakeLists.txt
@@ -3,13 +3,13 @@
 ##############################################################################
 
 cmake_minimum_required(VERSION 2.8.3)
-project(py_trees)
+project(py_trees_ros)
 
 ##############################################################################
 # Catkin
 ##############################################################################
 
-find_package(catkin REQUIRED COMPONENTS py_trees_msgs)
+find_package(catkin REQUIRED)
 
 catkin_python_setup()
 
@@ -20,21 +20,20 @@ catkin_package()
 # Installs
 ##############################################################################
 
-#install(FILES src/rocon_uri/rules/rules.yaml DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}/rules.yaml)
-
-catkin_install_python(
-    PROGRAMS
-        scripts/demo_behaviour
-        scripts/demo_dot_graphs
-        scripts/demo_tree
-    DESTINATION
-        ${CATKIN_PACKAGE_BIN_DESTINATION}
+ catkin_install_python(
+     PROGRAMS
+         scripts/demo_tree
+         scripts/blackboard_watcher
+     DESTINATION
+         ${CATKIN_PACKAGE_BIN_DESTINATION}
 )
+
+install(DIRECTORY launch DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION})
 
 ##############################################################################
 # Unit Tests
 ##############################################################################
 
 if (CATKIN_ENABLE_TESTING)
-  add_subdirectory(tests)
+  #add_subdirectory(tests)
 endif()
