+++ py_trees_ros/py_trees_ros/__init__.py
@@ -1,0 +1,23 @@
+#
+# License: BSD
+#   https://raw.github.com/stonier/py_trees_suite/license/LICENSE
+#
+##############################################################################
+# Documentation
+##############################################################################
+
+"""
+This is the top-level namespace of the py_trees_ros_
+package. It provides trees and behaviours to assist in
+robotics specific behaviour tree development.
+
+.. _py_trees_ros: http://wiki.ros.org/py_trees_ros
+"""
+
+##############################################################################
+# Imports
+##############################################################################
+
+from . import blackboard
+from . import subscribers
+from . import trees
