+++ py_trees_ros/scripts/demo_tree
@@ -17,6 +17,14 @@ Simple demo program for py trees
 
 import argparse
 import py_trees
+import py_trees.common
+import py_trees.composites
+import py_trees.behaviours
+import py_trees.display
+import py_trees.logging
+import py_trees_ros.trees
+import py_trees.trees
+import rospy
 import sys
 
 ##############################################################################
@@ -86,7 +94,7 @@ def render(root):
 if __name__ == '__main__':
     every_n_success = py_trees.behaviours.SuccessEveryN("EveryN", 5)
 
-    sequence = py_trees.Sequence(name="Sequence")
+    sequence = py_trees.composites.Sequence(name="Sequence")
     guard = py_trees.behaviours.Success("Guard")
     periodic_success = py_trees.behaviours.Periodic("Periodic", 3)
     finisher = py_trees.behaviours.Success("Finisher")
@@ -97,21 +105,27 @@ if __name__ == '__main__':
 
     idle = py_trees.behaviours.Success("Idle")
 
-    root = py_trees.Selector(name="Demo Tree")
+    root = py_trees.composites.Selector(name="Demo Tree")
     root.add_child(every_n_success)
     root.add_child(sequence)
     root.add_child(idle)
 
-    tree = py_trees.BehaviourTree(root)
-
-    parser = argparse.ArgumentParser(description='Demo the behaviour trees', usage=show_usage(tree.root))
+    parser = argparse.ArgumentParser(description='Demo the behaviour trees', usage=show_usage(root))
     parser.add_argument('-r', '--render', action='store_true', help='render dot tree to file.')
 
-    args = parser.parse_args(args=sys.argv[1:])
+    myargs = rospy.myargv(argv=sys.argv)
+
+    args = parser.parse_args(args=myargs[1:])
 
     if args.render:
         render(root)
         sys.exit()
 
+    # we've already imported rospy above, assume that we've got it
+    rospy.init_node("demo_tree")
+    # rebuild the tree as a ros behaviour tree
+    tree = py_trees_ros.trees.BehaviourTree(root)
+    tree.setup(15)
+
     tree.visitors.append(Visitor())
-    tree.tick_tock(sleep_ms=500, number_of_iterations=py_trees.CONTINUOUS_TICK_TOCK, pre_tick_handler=pre_tick_handler)
+    tree.tick_tock(sleep_ms=500, number_of_iterations=py_trees.trees.CONTINUOUS_TICK_TOCK, pre_tick_handler=pre_tick_handler)
