+++ py_trees_ros/tests/CMakeLists.txt
@@ -9,3 +9,9 @@
 ##############################################################################
 
 catkin_add_nosetests(nosetests)
+
+##############################################################################
+# Ros Tests
+##############################################################################
+
+add_subdirectory(rostests)
